them tinh nang check id co ton tai + nhac nho nguoi dung khi muon xoa 1 id

+++ src/assignment/java/controller/UserController.java
@@ -6,8 +6,12 @@
 package assignment.java.controller;
 
 import assignment.java.entity.User;
+import assignment.java.model.DAO;
 import assignment.java.model.UserModel;
 import assignment.java.utility.ScannerUtil;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 
 /**
  *
@@ -29,35 +33,80 @@ public class UserController {
         user.setId(id);
         user.setName(name);
         UserModel.insert(user);
+        UserModel.list();
     }
      
     public static void processUpdate() {
-        System.out.println("Thực hiện update sinh viên.");
-        System.out.println("Nhập id.");
-        int id = ScannerUtil.getInt();
-        System.out.println("Nhập name.");
-        String name = ScannerUtil.getString();
-        User user = new User();
-        user.setId(id);
-        user.setName(name);
-        UserModel.update(user);
+         try {
+           System.out.println("Thực hiện update sinh viên.");
+           System.out.println("Nhập id.");
+           User user = new User();
+           int id = ScannerUtil.getInt();
+           Statement statement = DAO.getConnection().createStatement();
+           String query = "select * from sv "
+                   + "where id='" + id + "'";
+           ResultSet rs = statement.executeQuery(query);
+           
+                if (!rs.next()) {
+                        System.out.println("The Id is not exist!");
+                } else {
+                        System.out.println("Danh sach ban muon Update:");
+                        user.setId(id);
+                        UserModel.select(user);
+                        System.out.println("Enter the Name you want to update:");
+                        String name = ScannerUtil.getString();
+                        user.setName(name);
+                        UserModel.update(user);
+                        
+                }  
+            
+
+        } catch (SQLException e) {
+            System.err.println("Lỗi khi select.");
+        }       
     }
     
     public static void processDelete() {
-        System.out.println("Thực hiện delete sinh viên.");
+        try {
+           System.out.println("Thực hiện delete sinh viên.");
+           System.out.println("Nhập id.");
+           User user = new User();
+           int id = ScannerUtil.getInt();
+           Statement statement = DAO.getConnection().createStatement();
+           String query = "select * from sv "
+                   + "where id='" + id + "'";
+           ResultSet rs = statement.executeQuery(query);
+           
+                if (!rs.next()) {
+                        System.out.println("The Id is not exist!");
+                } else{
+                        System.out.println("Danh sach ban muon Delete:");
+                        user.setId(id);
+                        UserModel.select(user);
+                        System.out.println("Are you sure to delete that?(y/n):");
+                        String yn = ScannerUtil.getString();
+                        if (yn.equals("y")) {
+                            UserModel.delete(user);
+                            UserModel.list();
+                        }else{
+                             System.out.println("Du lieu khong thay doi");
+                             UserModel.list();
+                        }                      
+                }  
+            
+
+        } catch (SQLException e) {
+            System.err.println("Lỗi khi select.");
+        }
+    }
+     
+    public static void processSelect() {
+        System.out.println("Thực hiện Select theo ID sinh viên.");
         System.out.println("Nhập id.");
         int id = ScannerUtil.getInt();
         User user = new User();
         user.setId(id);
-        UserModel.delete(user);
-    }
-     
-    public static void processSelect() {
-        System.out.println("Thực hiện Select theo tên sinh viên.");
-        System.out.println("Nhập name.");
-        String name = ScannerUtil.getString();
-        User user = new User();
-        user.setName(name);
         UserModel.select(user);
+        
     }
 }
